requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
pyrogram>=1.0.7
tgCrypto>=1.2.2
fontTools
gtts
tswift
pytz
youtube_dl
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
bing_image_downloader
google-api-python-client
pymongo==3.11.0
howdoi==2.0.10
gTTs-token
dnspython
cloudmersive-ocr-api-client==3.0.1
cloudmersive_virus_api_client==3.0.1
telegraph
dateparser==1.0.0
PyDictionary==2.0.1
better_profanity
nudepy
textblob 
nekos.py
opencv-python-headless==4.4.0.44
imutils==0.5.3
scikit-image==0.17.2
pornhub_api
youtube_search
search_engine_parser
asyncio
ffmpeg-python>=0.2.0
